Displaying appropriate colors when adding projects to the DOM.
Thinking about changing the color picker in the form to just a standard select-option just like how Todoist does. Also, will be moving on to the To-Dos and working with its logic.
Need to add a button for new To-Dos and display them after. Think of how you would like to display them onto the page.